requirements.txt
pandas 
numpy
matplotlib
seaborn
scipy
plotly
streamlit
langchain>=0.2
langchain-core>=0.2
python-dotenv>=1.0

langchain-openai>=0.2
langchain-google-genai>=1.0
langchain-ollama>=0.1
langgraph>=0.2
ipykernel
pytest>=8.0
